# JDBC: Optimize resolver for INSERT queries