Optimize `JdbcResolver`:
* Use `java.util.EnumSet` instead of switch-case sequence to define whether the provided `OneRow` field type is supported;
* Stop using `ThreadLocal` objects holding `SimpleDateFormat`, move the latter to instances of `JdbcResolver` instead;
* Remove type incompatibility warning log call (called when the provided field is of type TEXT, but table DDL states the field has other type, so `JdbcResolver` tries to make type conversion anyway).

All changes deal only with INSERT queries logic.


### Perfomance improvement
This patch was tested in the test case described below.

The perfomance improvement is as follows:

|  | [Before patch](https://github.com/greenplum-db/pxf/tree/9c28afa07a8e953dc5d54a13cfd4a7829e53f9e9) | [Before patch](https://github.com/greenplum-db/pxf/tree/9c28afa07a8e953dc5d54a13cfd4a7829e53f9e9) with log level `ERROR` | After patch |
| --- | --- | --- | --- |
| Execution time | 64±1 s | 29±0.5 s | 5.1±0.1 s |

#### Test case
For the test, the following table **DDL** was used:
```
CREATE WRITABLE EXTERNAL TABLE ch_t1_wpart (a INT, b INT, c TIMESTAMP)
LOCATION ('pxf://r1.t1?profile=jdbc&SERVER=ch&BATCH_SIZE=500000&POOL_SIZE=2')
FORMAT 'custom' (formatter='pxfwritable_export');
```
To our best knowledge, such DDL is the most effective one for our test case.

The dataset consisted of **five million rows** with all fields set to random values.

GPDB cluster consisted of **four segments** on two hosts. **Two PXF instances** were used (one on each GPDB host).

The host CPUs were `Intel(R) Xeon(R) CPU E5-4650 v2 @ 2.40GHz` (4 cores).

[Clickhouse](https://clickhouse.yandex) was used as an external database. It was run on two separate hosts, each connected to GPDB segment hosts with (reliable) 100Mbps connection.

Clickhouse table DDL (executed on each of two hosts):
```
create table r1.t1 (a UInt32 codec(ZSTD(1)), b UInt32 codec(ZSTD(1)), c DateTime codec(ZSTD(1))) 
engine ReplicatedMergeTree('/clickhouse/tables/s1/t1', 'r1') 
order by (a);
```